requirements.txt
aiofiles==0.8.0
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==21.4.0
charset-normalizer==2.1.1
colorlog==6.6.0
croniter==1.3.5
frozenlist==1.3.1
hikari==2.0.0.dev109
hikari-crescent==0.2.4
idna==3.3
multidict==6.0.2
mypy-extensions==0.4.3
pycooldown==0.1.0b8
python-dateutil==2.8.2
python-dotenv==0.20.0
six==1.16.0
types-croniter==1.3.2
yarl==1.8.1
